feat(fc-pallet-pass): add `SkipCheckIfPassAccount`
This new extension skips certain checks (like `CheckNonce`) if the caller is a session key.

This extension should be deprecated in the future in favour of a couple of `TransactionExtension`s that: 1) issues an origin for the account, and 2) after checks like Nonce, replaces this custom origin with `frame_system::Origin::signed(account_key)`, where `account_key` is the `AccountId` of the account being authorized by `pallet-pass`.